data_collection/carla/requirements.txt
numpy
pyopengl
glfw
imgui[glfw]
#git+https://github.com/pyimgui/pyimgui.git@dev/version-2.0[glfw]
opencv-python
jurigged
torch
torchvision
torchaudio
#carla
pyquaternion
tqdm
efficientnet_pytorch
matplotlib
nuscenes-devkit
open3d
dearpygui
transforms3d
numba
dill
docker
paramiko
